64th
Chromebook form correction on exceptions and number verifications

+++ ChromebookForm.cs
@@ -31,9 +31,9 @@ namespace Webber_Inventory_Search_2017_2018
             studentNameTextBox.Text = "";
             studentEditTextBox.Text = "";
             teacherNameTextBox.Text = "";
-            originalAddTextBox.Text = "";
+            originalAddTextBox1.Text = "";
             statusComboBox.Text = "";
-            loanAddTextBox.Text = "";
+            loanAddTextBox1.Text = "";
             status2ComboBox.Text = "";
             billAmountTextBox.Text = "";
             billDateTextBox.Text = "";
@@ -43,66 +43,90 @@ namespace Webber_Inventory_Search_2017_2018
         {
             try
             {
+                CheckDigitClass checkDigit = new CheckDigitClass();
+
+                if (loanAddTextBox.Text != "")
+                {
+                    bool loanDigit = checkDigit.digitOnly(loanAddTextBox.Text);
+                    if (loanDigit != true)
+                    {
+                        MessageBox.Show("Please enter numbers only for loan tag.");
+                        return;
+                    }
+                        
+                }
+                    
+
                 // Verify to see if all fields are entered
                 if (studentNameTextBox.Text != "" && studentEditTextBox.Text != "" && teacherNameTextBox.Text != "" && originalAddTextBox.Text != "" && statusComboBox.Text != "")
                 {
-                    string fullName = "";
-                    int lunchID = 0;
-                    string teacherName = "";
-                    int originalTag = 0;
-                    string status = "";
-                    int loanTag = 0;
-                    string status2 = "";
-                    float billAmount = 0;
-                    string billDate = "";
-                    string notes = "";
+                    bool originalDigit = checkDigit.digitOnly(originalAddTextBox.Text);
 
-                    fullName = studentNameTextBox.Text;
-                    lunchID = int.Parse(studentEditTextBox.Text);
-                    teacherName = teacherNameTextBox.Text;
-                    originalTag = int.Parse(originalAddTextBox.Text);
-                    status = statusComboBox.Text;
+                    if (originalDigit == true)
+                    {
+                        if (loanAddTextBox.Text != "")
+                        {
+                                string fullName = "";
+                                int lunchID = 0;
+                                string teacherName = "";
+                                int originalTag = 0;
+                                string status = "";
+                                int loanTag = 0;
+                                string status2 = "";
+                                float billAmount = 0;
+                                string billDate = "";
+                                string notes = "";
 
-                    if (loanAddTextBox.Text == "")
-                        loanTag = 0;
+                                fullName = studentNameTextBox.Text;
+                                lunchID = int.Parse(studentEditTextBox.Text);
+                                teacherName = teacherNameTextBox.Text;
+                                originalTag = int.Parse(originalAddTextBox1.Text);
+                                status = statusComboBox.Text;
+
+                                if (loanAddTextBox1.Text == "")
+                                    loanTag = 0;
+                                else
+                                    loanTag = int.Parse(loanAddTextBox1.Text);
+
+                                if (status2ComboBox.Text == "")
+                                    status2 = "";
+                                else
+                                    status2 = status2ComboBox.Text;
+
+                                if (billAmountTextBox.Text == "")
+                                    billAmount = 0;
+                                else
+                                    billAmount = float.Parse(billAmountTextBox.Text);
+
+                                if (billDateTextBox.Text == "")
+                                    billDate = "";
+                                else
+                                    billDate = billDateTextBox.Text;
+
+                                if (textBox1.Text == "")
+                                    notes = "";
+                                else
+                                    notes = textBox1.Text;
+
+                                // Open connection to database
+                                connection.Open();
+
+                                OleDbCommand command = new OleDbCommand();
+                                command.Connection = connection;
+                                command.CommandText = "insert into Chromebook_Information (FullName,LunchID,TeacherName,OriginalTag,Status,LoanTag,LoanStatus,BillAmount,BillDate,Notes) values('" + fullName + "'," + lunchID + ",'" + teacherName + "'," + originalTag + ",'" + status + "'," + loanTag + ",'" + status2 + "'," + billAmount + ",'" + billDate + "','" + notes + "')"; //,'" + note + "'
+                                command.ExecuteNonQuery();
+                                connection.Close();
+
+                                MessageBox.Show("Chromebook was successfully added!");
+
+                                clearAll();
+                            }
+                    }
                     else
-                        loanTag = int.Parse(loanAddTextBox.Text);
-
-                    if (status2ComboBox.Text == "")
-                        status2 = "";
-                    else
-                        status2 = status2ComboBox.Text;
-
-                    if (billAmountTextBox.Text == "")
-                        billAmount = 0;
-                    else
-                        billAmount = float.Parse(billAmountTextBox.Text);
-
-                    if (billDateTextBox.Text == "")
-                        billDate = "";
-                    else
-                        billDate = billDateTextBox.Text;
-
-                    if (textBox1.Text == "")
-                        notes = "";
-                    else
-                        notes = textBox1.Text;
-
-                    // Open connection to database
-                    connection.Open();
-
-                    OleDbCommand command = new OleDbCommand();
-                    command.Connection = connection;
-                    command.CommandText = "insert into Chromebook_Information (FullName,LunchID,TeacherName,OriginalTag,Status,LoanTag,LoanStatus,BillAmount,BillDate,Notes) values('" + fullName + "'," + lunchID + ",'" + teacherName + "'," + originalTag + ",'" + status + "'," + loanTag + ",'" + status2 + "'," + billAmount + ",'" + billDate + "','" + notes + "')"; //,'" + note + "'
-                    command.ExecuteNonQuery();
-                    connection.Close();
-
-                    MessageBox.Show("Chromebook was successfully added!");
-
-                    clearAll();
-                }
+                        MessageBox.Show("Please enter numbers only for original tag.");
+                }    
                 else
-                    MessageBox.Show("Fill out all required fields."); 
+                    MessageBox.Show("Please fill out all required fields."); 
             }
             catch (Exception)
             {
@@ -121,10 +145,10 @@ namespace Webber_Inventory_Search_2017_2018
             try
             {
                 // Verify to see if all fields are entered
-                if (studentEditTextBox.Text != "" && originalAddTextBox.Text != "" && statusComboBox.Text != "")
+                if (studentEditTextBox.Text != "" && originalAddTextBox1.Text != "" && statusComboBox.Text != "")
                 {
                     int lunchID = int.Parse(studentEditTextBox.Text);
-                    int originalTag = int.Parse(originalAddTextBox.Text);
+                    int originalTag = int.Parse(originalAddTextBox1.Text);
                     string status = statusComboBox.Text;
                     int loanTag;
                     string status2;
@@ -132,10 +156,10 @@ namespace Webber_Inventory_Search_2017_2018
                     string billDate;
                     string notes;
 
-                    if (loanAddTextBox.Text == "")
+                    if (loanAddTextBox1.Text == "")
                         loanTag = 0;
                     else
-                        loanTag = int.Parse(loanAddTextBox.Text);
+                        loanTag = int.Parse(loanAddTextBox1.Text);
 
                     if (status2ComboBox.Text == "")
                         status2 = "";
@@ -182,26 +206,40 @@ namespace Webber_Inventory_Search_2017_2018
 
         private void deleteButton_Click(object sender, EventArgs e)
         {
-            try
+            if (deleteTextBox.Text == "")
             {
-                // Open database and delete all data for selected item
-                connection.Open();
-
-                int deleteItem = int.Parse(deleteTextBox.Text);
-
-                OleDbCommand command = new OleDbCommand();
-                command.Connection = connection;
-                command.CommandText = "delete from Chromebook_Information where ID=" + deleteItem + "";
-                command.ExecuteNonQuery();
-                connection.Close();
-
-                MessageBox.Show("All data was deleted for Chromebook #" + deleteTextBox.Text + ".");
-                deleteTextBox.Text = "";
+                MessageBox.Show("Please enter a number.");
             }
-            catch (Exception)
+            else
             {
-                MessageBox.Show("Oops, something went wrong. Please exit and try again.");
-            }
+                CheckDigitClass checkDigit = new CheckDigitClass();
+                bool digit = checkDigit.digitOnly(deleteTextBox.Text);
+                if (digit == true)
+                {
+                    try
+                    {
+                        // Open database and delete all data for selected item
+                        connection.Open();
+
+                        int deleteItem = int.Parse(deleteTextBox.Text);
+
+                        OleDbCommand command = new OleDbCommand();
+                        command.Connection = connection;
+                        command.CommandText = "delete from Chromebook_Information where ID=" + deleteItem + "";
+                        command.ExecuteNonQuery();
+                        connection.Close();
+
+                        MessageBox.Show("All data was deleted for Chromebook #" + deleteTextBox.Text + ".");
+                        deleteTextBox.Text = "";
+                    }
+                    catch (Exception)
+                    {
+                        MessageBox.Show("Oops, something went wrong. Please exit and try again.");
+                    }
+                }
+                else
+                    MessageBox.Show("Please input numbers only.");
+            }     
         }
 
         private void searchChromebookButton_Click(object sender, EventArgs e)
@@ -294,37 +332,58 @@ namespace Webber_Inventory_Search_2017_2018
 
         private void getInfoButton_Click(object sender, EventArgs e)
         {
-            int id = int.Parse(idTextBox.Text);
-            // Check for which brand is selected
-            try
+            if (idTextBox.Text == "")
             {
-                connection.Open();
-
-                OleDbCommand command = new OleDbCommand();
-                command.Connection = connection;
-                string query = "select * from Chromebook_Information where ID=" + id + "";
-                command.CommandText = query;
-
-                OleDbDataReader reader = command.ExecuteReader();
-                while (reader.Read())
-                {
-                    studentNameTextBox.Text = reader["FullName"].ToString();
-                    studentEditTextBox.Text = reader["LunchID"].ToString();
-                    teacherNameTextBox.Text = reader["TeacherName"].ToString();
-                    originalAddTextBox.Text = reader["OriginalTag"].ToString();
-                    statusComboBox.Text = reader["Status"].ToString();
-                    loanAddTextBox.Text = reader["LoanTag"].ToString();
-                    status2ComboBox.Text = reader["LoanStatus"].ToString();
-                    billAmountTextBox.Text = reader["BillAmount"].ToString();
-                    billDateTextBox.Text = reader["BillDate"].ToString();
-                    textBox1.Text = reader["Notes"].ToString();
-                }
-
-                connection.Close();
+                MessageBox.Show("Please enter a number.");
             }
-            catch (Exception ex)
+            else
             {
-                MessageBox.Show("Please go back and try again." + ex);
+                CheckDigitClass checkDigit = new CheckDigitClass();
+                bool idDigit = checkDigit.digitOnly(idTextBox.Text);
+                if (idDigit == true)
+                {
+                    // Convert string to int
+                    int id = int.Parse(idTextBox.Text);
+
+                    try
+                    {
+                        connection.Open();
+
+                        OleDbCommand command = new OleDbCommand();
+                        command.Connection = connection;
+                        string query = "select * from Chromebook_Information where ID=" + id + "";
+                        command.CommandText = query;
+
+                        OleDbDataReader reader = command.ExecuteReader();
+                        while (reader.Read())
+                        {
+                            studentNameTextBox.Text = reader["FullName"].ToString();
+                            studentEditTextBox.Text = reader["LunchID"].ToString();
+                            teacherNameTextBox.Text = reader["TeacherName"].ToString();
+                            originalAddTextBox1.Text = reader["OriginalTag"].ToString();
+                            statusComboBox.Text = reader["Status"].ToString();
+                            loanAddTextBox1.Text = reader["LoanTag"].ToString();
+                            status2ComboBox.Text = reader["LoanStatus"].ToString();
+                            billAmountTextBox.Text = reader["BillAmount"].ToString();
+                            billDateTextBox.Text = reader["BillDate"].ToString();
+                            textBox1.Text = reader["Notes"].ToString();
+                        }
+
+                        connection.Close();
+                    }
+                    catch (Exception ex)
+                    {
+                        string page = "Chromebook";
+                        string button = "Quick Fill";
+                        string exception = ex.ToString();
+                        BugSplatForm bugSplat = new BugSplatForm(page, button, exception);
+                        bugSplat.ShowDialog();
+
+                        this.Close();
+                    }
+                }
+                else
+                    MessageBox.Show("Please enter numbers only.");
             }
         }
     }
